remove Factory
