+++ pkg/externalresource/broker/broker.go
@@ -2,7 +2,10 @@ package broker
 
 import (
 	"context"
+	"path/filepath"
 
+	"github.com/gogo/status"
+	"github.com/pingcap/errors"
 	"github.com/pingcap/tiflow/dm/pkg/log"
 	"go.uber.org/zap"
 
@@ -15,8 +18,9 @@ import (
 type DefaultBroker struct {
 	config     *storagecfg.Config
 	executorID resModel.ExecutorID
+	client     *rpcutil.FailoverRPCClients[pb.ResourceManagerClient]
 
-	factory *Factory
+	fileManager FileManager
 }
 
 func NewBroker(
@@ -24,18 +28,16 @@ func NewBroker(
 	executorID resModel.ExecutorID,
 	client *rpcutil.FailoverRPCClients[pb.ResourceManagerClient],
 ) *DefaultBroker {
+	fm := NewLocalFileManager(*config.Local)
 	return &DefaultBroker{
-		config:     config,
-		executorID: executorID,
-		factory: &Factory{
-			config:     config,
-			client:     client,
-			executorID: executorID,
-		},
+		config:      config,
+		executorID:  executorID,
+		client:      client,
+		fileManager: fm,
 	}
 }
 
-func (i *DefaultBroker) OpenStorage(
+func (b *DefaultBroker) OpenStorage(
 	ctx context.Context,
 	workerID resModel.WorkerID,
 	jobID resModel.JobID,
@@ -48,7 +50,7 @@ func (i *DefaultBroker) OpenStorage(
 
 	switch tp {
 	case resModel.ResourceTypeLocalFile:
-		return i.factory.NewHandleForLocalFile(ctx, jobID, workerID, resourcePath)
+		return b.newHandleForLocalFile(ctx, jobID, workerID, resourcePath)
 	case resModel.ResourceTypeS3:
 		log.L().Panic("resource type s3 is not supported for now")
 	default:
@@ -58,6 +60,110 @@ func (i *DefaultBroker) OpenStorage(
 	panic("unreachable")
 }
 
-func (i *DefaultBroker) OnWorkerClosed(ctx context.Context, workerID resModel.WorkerID, jobID resModel.JobID) {
-	panic("implement me")
+func (b *DefaultBroker) OnWorkerClosed(ctx context.Context, workerID resModel.WorkerID, jobID resModel.JobID) {
+	err := b.fileManager.RemoveTemporaryFiles(workerID)
+	if err != nil {
+		// TODO when we have a cloud-based error collection service, we need
+		// to report this.
+		// However, since an error here is unlikely to indicate a correctness
+		// problem, we do not take further actions.
+		log.L().Warn("Failed to remove temporary files for worker",
+			zap.String("worker-id", workerID),
+			zap.String("job-id", jobID),
+			zap.Error(err))
+	}
+}
+
+func (b *DefaultBroker) newHandleForLocalFile(
+	ctx context.Context,
+	jobID resModel.JobID,
+	workerID resModel.WorkerID,
+	resourceID resModel.ResourceID,
+) (Handle, error) {
+	tp, suffix, err := resModel.ParseResourcePath(resourceID)
+	if err != nil {
+		return nil, err
+	}
+	if tp != resModel.ResourceTypeLocalFile {
+		log.L().Panic("unexpected resource type", zap.String("type", string(tp)))
+	}
+
+	record, exists, err := b.checkForExistingResource(ctx, resourceID)
+	if err != nil {
+		return nil, err
+	}
+
+	var creatorWorkerID string
+	if exists {
+		creatorWorkerID = record.Worker
+	} else {
+		creatorWorkerID = workerID
+	}
+	filePath := filepath.Join(getWorkerDir(b.config, creatorWorkerID), suffix)
+	log.L().Info("Using local storage with path", zap.String("path", filePath))
+
+	ls, err := newBrStorageForLocalFile(filePath)
+	if err != nil {
+		return nil, err
+	}
+
+	return &BrExternalStorageHandle{
+		inner:  ls,
+		client: b.client,
+
+		id:         resourceID,
+		jobID:      jobID,
+		workerID:   creatorWorkerID,
+		executorID: b.executorID,
+	}, nil
+}
+
+func (b *DefaultBroker) checkForExistingResource(
+	ctx context.Context,
+	resourceID resModel.ResourceID,
+) (*resModel.ResourceMeta, bool, error) {
+	resp, err := rpcutil.DoFailoverRPC(
+		ctx,
+		b.client,
+		&pb.QueryResourceRequest{ResourceId: resourceID},
+		pb.ResourceManagerClient.QueryResource,
+	)
+	if err == nil {
+		return &resModel.ResourceMeta{
+			ID:       resourceID,
+			Job:      resp.GetJobId(),
+			Worker:   resp.GetCreatorWorkerId(),
+			Executor: resModel.ExecutorID(resp.GetCreatorExecutor()),
+			Deleted:  false,
+		}, true, nil
+	}
+
+	// TODO perhaps we need a grpcutil package to put all this stuff?
+	st, ok := status.FromError(err)
+	if !ok {
+		// If the error is not derived from a grpc status, we should throw it.
+		return nil, false, errors.Trace(err)
+	}
+	if len(st.Details()) != 1 {
+		// The resource manager only generates status with ONE detail.
+		return nil, false, errors.Trace(err)
+	}
+	resourceErr, ok := st.Details()[0].(*pb.ResourceError)
+	if !ok {
+		return nil, false, errors.Trace(err)
+	}
+
+	log.L().Info("Got ResourceError",
+		zap.String("resource-id", resourceID),
+		zap.Any("resource-err", resourceErr))
+	switch resourceErr.ErrorCode {
+	case pb.ResourceErrorCode_ResourceNotFound:
+		// Indicates that there is no existing resource with the same name.
+		return nil, false, nil
+	default:
+		log.L().Warn("Unexpected ResourceError",
+			zap.String("code", resourceErr.ErrorCode.String()),
+			zap.String("stack-trace", resourceErr.StackTrace))
+		return nil, false, errors.Trace(err)
+	}
 }
